Renaming/reorganizing files, modules, and test data; moving non-upload journal operations to separate module
