+++ chorus_upload_journal/__init__.py
@@ -1,0 +1,3 @@
+"""Contains modules for creating and working with CHoRUS upload journals"""
+
+__version__ = "0.1"
