+++ tests/README.md
@@ -1,0 +1,21 @@
+# create_test_data
+Creates a folder structure and fake data files for each of the three modalities (OMOP, images, and waveforms).
+
+
+# INSTALLATION
+
+First install the python package "flit"
+
+```
+pip install flit
+```
+
+## Local Dev Installation
+
+This will set up a symlink in the site-package directory to the source tree.
+
+From the root of the source tree, run
+
+```
+flit install --symlink
+```
