+++ tests/create_test_data/__init__.py
@@ -1,0 +1,5 @@
+"""
+Create a synthetic folder structure for testing of journaling functionality.
+"""
+
+__version__ = "0.1"
