+++ tests/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["flit_core >=3.2,<4"]
 build-backend = "flit_core.buildapi"
 
 [project]
-name = "chorus_upload_journal"
+name = "create_test_data"
 authors = [
     {name = "Will Ashe"},
     {name = "Tony Pan"},
@@ -15,14 +15,7 @@ dynamic = ["version", "description"]
 requires-python = ">=3.10"
 dependencies = [
     "numpy",
-    "pandas",
-    "cloudpathlib",
-    "cloudpathlib[azure]",
-    "tomli"
 ]
 
 [project.optional-dependencies]
-s3 = ["cloudpathlib[s3]",
-]
-gs = ["cloudpathlib[gs]",
-]
+
