--- upload_manifest/__init__.py
@@ -1,3 +1,0 @@
-"""An amazing sample package!"""
-
-__version__ = "0.1"
